+++ json2yolo/general_json2yolo.py
@@ -4,7 +4,7 @@ import cv2
 import pandas as pd
 from PIL import Image
 
-from utils import *
+from json2yolo.utils import *
 
 
 # Convert INFOLKS JSON file into YOLO-format labels ----------------------------
@@ -295,78 +295,12 @@ def convert_coco_json(json_dir='../coco/annotations/', use_segments=False, cls91
                 with open((fn / f).with_suffix('.txt'), 'a') as file:
                     file.write(('%g ' * len(line)).rstrip() % line + '\n')
 
-def min_index(arr1, arr2):
-    """Find a pair of indexes with the shortest distance. 
-    Args:
-        arr1: (N, 2).
-        arr2: (M, 2).
-    Return:
-        a pair of indexes(tuple).
-    """
-    # (N, M)
-    dis = ((arr1[:, None, :] - arr2[None, :, :]) ** 2).sum(-1)
-    index = np.unravel_index(np.argmin(dis, axis=None), dis.shape)
-    return index
-
-
-def merge_multi_segment(segments):
-    """Merge multi segments to one list.
-    Find the coordinates with min distance between each segment,
-    then connect these coordinates with one thin line to merge all 
-    segments into one.
-
-    Args:
-        segments(List(List)): original segmentations in coco's json file.
-            like [segmentation1, segmentation2,...], 
-            each segmentation is a list of coordinates.
-    """
-    s = []
-    segments = [np.array(i).reshape(-1, 2) for i in segments]
-    idx_list = [[] for _ in range(len(segments))]
-
-    # record the indexes with min distance between each segment
-    for i in range(1, len(segments)):
-        idx1, idx2 = min_index(segments[i - 1], segments[i])
-        idx_list[i - 1].append(idx1)
-        idx_list[i].append(idx2)
-
-    # use two round to connect all the segments
-    for k in range(2):
-        # forward connection
-        if k == 0:
-            for i, idx in enumerate(idx_list):
-                # middle segments have two indexes
-                # reverse the index of middle segments
-                if len(idx) == 2 and idx[0] > idx[1]:
-                    idx = idx[::-1]       
-                    segments[i] = segments[i][::-1, :]
-
-                segments[i] = np.roll(segments[i], -idx[0], axis=0)
-                segments[i] = np.concatenate([segments[i], segments[i][0:1]])
-                # deal with the first segment and the last one
-                if i == 0 or i == (len(idx_list) - 1):
-                    s.append(segments[i])
-                # deal with the middle ones
-                else:
-                    idx = [0, idx[1] - idx[0]]
-                    s.append(segments[i][idx[0]:idx[1] + 1])
-
-        # backward connection
-        else:
-            for i in range(len(idx_list) - 1, -1, -1):
-                if i != 0 and i != (len(idx_list) - 1):
-                    idx = idx_list[i]
-                    nidx = abs(idx[1] - idx[0])
-                    s.append(segments[i][nidx:])
-    return s
-
-
 
 if __name__ == '__main__':
     source = 'COCO'
 
     if source == 'COCO':
-        convert_coco_json('../../Downloads/Objects365')  # directory with *.json
+        convert_coco_json('data/train')  # directory with *.json
 
     elif source == 'infolks':  # Infolks https://infolks.info/
         convert_infolks_json(name='out',
